+++ src/states/common/base.rs
@@ -6,9 +6,9 @@
 // KIND, either express or implied. Please review the Licences for the specific language governing
 // permissions and limitations relating to use of the SAFE Network Software.
 
-use super::Core;
 use crate::{
     action::Action,
+    core::Core,
     error::{Result, RoutingError},
     id::PublicId,
     location::{DstLocation, SrcLocation},
