+++ src/states/common/mod.rs
@@ -9,85 +9,7 @@
 mod base;
 
 pub use self::base::Base;
-use crate::{
-    id::FullId,
-    location::DstLocation,
-    message_filter::MessageFilter,
-    messages::{Message, Variant},
-    quic_p2p::Token,
-    rng::MainRng,
-    time::Duration,
-    timer::Timer,
-    transport::{PeerStatus, Transport},
-    xor_space::XorName,
-};
-use bytes::Bytes;
-use std::{net::SocketAddr, slice};
+use crate::time::Duration;
 
 /// Delay after which a bounced message is resent.
 pub const BOUNCE_RESEND_DELAY: Duration = Duration::from_secs(1);
-
-/// Struct that contains data common to all states.
-pub struct Core {
-    pub full_id: FullId,
-    pub transport: Transport,
-    pub msg_filter: MessageFilter,
-    pub timer: Timer,
-    pub rng: MainRng,
-}
-
-impl Core {
-    pub fn name(&self) -> &XorName {
-        self.full_id.public_id().name()
-    }
-
-    pub fn send_message_to_targets(
-        &mut self,
-        conn_infos: &[SocketAddr],
-        dg_size: usize,
-        msg: Bytes,
-    ) {
-        self.transport
-            .send_message_to_targets(conn_infos, dg_size, msg)
-    }
-
-    pub fn send_message_to_target_later(
-        &mut self,
-        dst: &SocketAddr,
-        message: Bytes,
-        delay: Duration,
-    ) {
-        self.transport
-            .send_message_to_target_later(dst, message, &self.timer, delay)
-    }
-
-    pub fn send_direct_message(&mut self, recipient: &SocketAddr, variant: Variant) {
-        let message = match Message::single_src(&self.full_id, DstLocation::Direct, variant) {
-            Ok(message) => message,
-            Err(error) => {
-                error!("Failed to create message: {:?}", error);
-                return;
-            }
-        };
-
-        let bytes = match message.to_bytes() {
-            Ok(bytes) => bytes,
-            Err(error) => {
-                error!("Failed to serialize message {:?}: {:?}", message, error);
-                return;
-            }
-        };
-
-        self.send_message_to_targets(slice::from_ref(recipient), 1, bytes)
-    }
-
-    pub fn handle_unsent_message(
-        &mut self,
-        addr: SocketAddr,
-        msg: Bytes,
-        msg_token: Token,
-    ) -> PeerStatus {
-        self.transport
-            .target_failed(msg, msg_token, addr, &self.timer)
-    }
-}
