+++ src/states/mod.rs
@@ -11,7 +11,7 @@ mod joining_peer;
 
 pub mod common;
 
-pub use self::{approved_peer::ApprovedPeer, common::Core, joining_peer::JoiningPeer};
+pub use self::{approved_peer::ApprovedPeer, joining_peer::JoiningPeer};
 
 // # The state machine
 //
